fix: M-mode GUI freeze during MP4/cine playback
Root cause: mmode_line_completed emitted on every mouse-move during
node drag, triggering recalculate_from_frames (all cached frames) in
GUI thread. For 300-frame cine this caused ~720MB allocations per
mouse-move → freeze → OOM kill.

Changes:
- Remove emit during drag, only emit on drag end (viewer_widget.py)
- Replace float64 with float32 in mmode_extractor (50% memory saving)
- Set autoLevels=False with fixed levels (0,255) in mmode_widget

+++ src/echo_personal_tool/domain/services/mmode_extractor.py
@@ -18,7 +18,7 @@ def extract_mmode_column(
     ys = start[1] + t * (end[1] - start[1])
     coords = np.array([ys, xs])
     result = map_coordinates(
-        frame.astype(np.float64),
+        frame.astype(np.float32),
         coords,
         order=1,
         mode="nearest",
